refac: Repository 패턴 적용, 예약 존재하는 테마 삭제 예외처리 추가

+++ src/main/java/nextstep/web/service/ThemeService.java
@@ -2,9 +2,11 @@ package nextstep.web.service;
 
 import java.util.List;
 
+import nextstep.domain.Reservation;
 import nextstep.domain.Theme;
 import nextstep.web.exceptions.ErrorCode;
 import nextstep.web.exceptions.ThemeException;
+import nextstep.web.repository.ReservationDAOImpl;
 import nextstep.web.repository.ThemeDAOImpl;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
@@ -12,9 +14,11 @@ import org.springframework.stereotype.Service;
 @Service
 public class ThemeService {
     private final ThemeDAOImpl themeDAOImpl;
+    private final ReservationDAOImpl reservationDAOImpl;
 
-    public ThemeService(ThemeDAOImpl themeDAOImpl) {
+    public ThemeService(ThemeDAOImpl themeDAOImpl, ReservationDAOImpl reservationDAOImpl) {
         this.themeDAOImpl = themeDAOImpl;
+        this.reservationDAOImpl = reservationDAOImpl;
     }
 
     public Long createTheme(Theme theme) {
@@ -31,12 +35,18 @@ public class ThemeService {
     }
 
     public Theme deleteTheme(Long id) {
+        Theme theme;
         try {
-            Theme theme = themeDAOImpl.findById(id);
-            themeDAOImpl.delete(id);
-            return theme;
+            theme = themeDAOImpl.findById(id);
         } catch(EmptyResultDataAccessException e) {
             throw new ThemeException(ErrorCode.THEME_NOT_FOUND);
         }
+
+        List<Reservation> reservations = reservationDAOImpl.findByTheme(theme);
+        if (reservations.size() > 0) {
+            throw new ThemeException(ErrorCode.RESERVATION_WITH_THIS_THEME_EXISTS);
+        }
+        themeDAOImpl.delete(id);
+        return theme;
     }
 }
